add centralized error handler and some field validation to post route

+++ routes/book_routes.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, request, jsonify
 from models.book import Book, db
+from errors.exceptions import APIError, BookNotFound
 
 book_bp = Blueprint("books", __name__)
 
@@ -7,21 +8,33 @@ book_bp = Blueprint("books", __name__)
 @book_bp.route("/books", methods=["GET"])
 def get_books():
     books = Book.query.all()
-    return jsonify([book.to_dict() for book in books])
+    return jsonify({ "success": True, "count": len(books), "data": [book.to_dict() for book in books] })
 
 
 # GET SINGLE BOOK
 @book_bp.route("/books/<int:id>", methods=["GET"])
 def get_book(id):
-    book = Book.query.get_or_404(id)
-    return jsonify(book.to_dict())
+    book = Book.query.get(id)
+    if not book:
+        raise BookNotFound()
+    return jsonify({ "success": True, "data": book.to_dict() })
 
 
 # CREATE BOOK
 @book_bp.route("/books", methods=["POST"])
 def add_book():
 
+    required_fields = ["title", "genre", "rating", "comment"]
+
     data = request.json
+
+    # Adding some request validation
+    if not data:
+        raise APIError(message="Request body must be JSON.", status_code=400)
+    
+    for field in required_fields:
+        if field not in data:
+            raise APIError(message=f"Field '{field}' is required.", status_code=400)
 
     book = Book(
         title=data["title"],
@@ -33,14 +46,16 @@ def add_book():
     db.session.add(book)
     db.session.commit()
 
-    return jsonify(book.to_dict()), 201
+    return jsonify({ "success": True, "data": book.to_dict() }), 201
 
 
 # UPDATE BOOK
 @book_bp.route("/books/<int:id>", methods=["PUT"])
 def update_book(id):
 
-    book = Book.query.get_or_404(id)
+    book = Book.query.get(id)
+    if not book:
+        raise BookNotFound()
     data = request.json
 
     book.title = data.get("name", book.title)
@@ -50,18 +65,21 @@ def update_book(id):
 
     db.session.commit()
 
-    return jsonify(book.to_dict())
+    return jsonify({ "success": True, "data": book.to_dict() })
 
 
 # DELETE BOOK
 @book_bp.route("/books/<int:id>", methods=["DELETE"])
 def delete_book(id):
 
-    book = Book.query.get_or_404(id)
+    book = Book.query.get(id)
+    if not book:
+        raise BookNotFound()
 
     db.session.delete(book)
     db.session.commit()
 
     return jsonify({
+        "success": True,
         "message": "Book deleted successfully."
     })
